refactor: 날짜 data 처리 위치 변경

+++ Diary/Model/Json/SampleDiaryContent.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct SampleDiaryContent: Decodable {
     let title, body: String
-    let createdAt: Int
+    let createdAt: Double
 
     enum CodingKeys: String, CodingKey {
         case title, body
